backend/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
